Update patch set 1
Patch Set 1:

> Hi Hughes, thanks a lot.
 > Recently, we meet a problem, the device with Android kitkat OS
 > can't pass the test case of statvfs in cts r3 which use rootfs "/"
 > as a check point. Finally, I found when our rootfs mounted with
 > romfs, the case passed, if mounted with tmpfs, it is failed. it
 > look like the old test case is not applied to the tmpfs. I want to
 > confirm if that is another reason for changing the check point from
 > "/" to "/proc". I known I can back port this patch to fix it
 > internal. but, I want to confirm this issue and if it will be fixed
 > in next cts.

yes, it was this kind of unpredictability that caused us to use /proc. ("Using /proc seems like a better idea because it's more similar across devices.") future versions of CTS will include this updated test. if you're having trouble passing CTS because of this, ask your Google contact to talk to me about either a waiver or backporting this fix to older CTSes.

Patch-set: 1