Update patch set 1
Patch Set 1:

or, indeed, whether we want to add this at all. i feel like this stuff keeps cropping up, and it's probably easier than trying to upstream changes to get people off *_unlocked and friends.

i'd be curious to see how much difference it makes to the runtime of, say, strace to use the unlocked functions.

Patch-set: 1