Update patch set 2
Patch Set 2:

Alternatives:
http://man7.org/linux/man-pages/man3/malloc_get_state.3.html
http://man7.org/linux/man-pages/man3/malloc_stats.3.html

These are both much better for the problem we're trying to solve than the XML one (since no one wants to have to add an xml parser to their library just for the sake of getting malloc stats). malloc_stats(3) would be essentially useless unless we send stderr to logcat, but that might not be a bad idea. Looking at the docs for malloc_get_state(3), I don't think that's actually an intended use of this function (because there's a malloc_set_state(3)).

Patch-set: 2
Label: Verified=0