Update patch set 1
Patch Set 1:

> It's a little weird to have this but exclude two of the functions.
 > It's possible we would see some fallout from projects trying to use
 > this but still needing to have android specific stuff to avoid
 > calling those functions.

yeah, it's annoying but i don't have a solution right now. (i really wish we'd made FILE opaque for LP64 so we could just replace the BSD implementation at some point.)

and i feel like half an eye is almost always better than no eye at all. in this case, for example, this is enough to build the rest of elfutils. (in fact, we don't even need all these functions for that. so there's a reasonable argument to implement less than we do here.)

i'll raise another bug for the rest of the *_unlocked stdio functions. they seem to be used quite a lot in linux land. even if any given project only uses one or two (which is how we came to have the random subset we currently have). strace, for example, only uses fputs_unlocked (though that's hidden behind configure so we're getting away with it).

Patch-set: 1
Label: Verified=0