Update patch set 2
Patch Set 2:

> Alternatives:
 > http://man7.org/linux/man-pages/man3/malloc_get_state.3.html
 > http://man7.org/linux/man-pages/man3/malloc_stats.3.html
 > 
 > These are both much better for the problem we're trying to solve
 > than the XML one (since no one wants to have to add an xml parser
 > to their library just for the sake of getting malloc stats).

no, the obvious users of this are DDMS and procrank, and they're both Java. so for them, reading XML is easier than dealing with some random format we make up ourselves. if we're not aiming that high right now then...

 > malloc_stats(3) would be essentially useless unless we send stderr
 > to logcat, but that might not be a bad idea.

...this is more like what we want. something an app can call that just does everything.

and the great thing about these is that they're somewhat portable and it's just us following existing practice. kind of like POSIX. which means we don't have to worry too much about what we're doing or how public it's going to be.

(obviously there are still interesting questions here since afaik neither format is actually defined. can we output compatible XML? the example i saw looked like we might be able to, which would be awesome.)

 > Looking at the docs
 > for malloc_get_state(3), I don't think that's actually an intended
 > use of this function (because there's a malloc_set_state(3)).

yeah, i was misled by that name too.

Patch-set: 2
Label: Verified=0