bionic: get correct stack size to avoid the out of stack range access.
On some arch (like IA and arm64), before transfer the execution to
user interpreter, kernel could adjust the stack pointer value with
random value between 0 and 8k if ALSR is enabled (kernel source in
create_elf_tables->arch_align_stack).

For main thread, the stack_top gotten based on current sp is not
correct. The distance between stack_top from sp and the real stack
top is random value between 0 and 8k.

So the stack_bottom is not correct as well. The worst case is that
it's not in real stack range.

ART hit this issue when they tried to add ImplicitProtection.

This change check the /proc/self/maps to find the real stack top
and shrink stack size to make sure stack_bottom is in stack range.

Signed-off-by: Fengwei Yin <fengwei.yin@intel.com>
Change-Id: I1d4dbffe7bc7bda1d353c3a295dbf68d29f63158