Update patch set 1
Patch Set 1:

Always got "500 internal server error" when I did inline comments. So I put my comments here.

For "c > end" from Hans, because we didn't clean buf, it could have the data parsed already. The memchr could get old data. The "c > end" here is to detect the last line was handled already and we didn't find what we need.

And all the 4096 will be changed to PAGE_SIZE for the patchset 2.

For the comment from Narayan, Not just IA has this behavior. We need this change if the arch does adjust the sp with random value.  The IA, arm64 and mips all adjust it. So we need this change for IA, arm64 and mips. Just arm does not need this change.

Patch-set: 1
Reviewer: Gerrit User 1011068 <1011068@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0