Update patch set 2
Patch Set 2:

> Please take a look of patch set 2.
 > Maybe that will be acceptable to apply to upstream source?

well, i think if we wanted to talk to upstream, we'd probably ask "hey, could you wrap your uses of __weak_alias in an #ifdef like the other BSDs do?".

 > I don't want to change the definition of __weak_alias too.
 > It might be okay to change __weak_alias for these 8 files, but not
 > for other files.

if you search you can see that those are the only upstream-openbsd files [that we import] that use __weak_alias. there's a lot more use in the netbsd code, but we already have a hack to #undef __weak_alias when building all that code.

Patch-set: 2