Update patch set 3
Patch Set 3:

Interesting... As far as I understand the static linker should take care of these things - in the form of omitting relocation for the referenced symbol. At least this is how it is done for protected onces. Let me see If I can write a test that produces relocation and is DT_SYMBOLIC.

Patch-set: 3