[x86,x86_64] Fix libgcc unwinding through signal
This change provides __restore() on x86 and __restore_rt() on x86_64 with
unwinding information to be able to unwind through signal frame via
libgcc provided unwinding interface. See comments inlined for more
details.

Also change test to remove dependency on
__attribute__((cleanup(foo_cleanup))) feature. It's not reliable, at
least in this particular case. I've got a case when adding irrelevant
printf in different places makes test pass or fail.
Instead just count the frames we are looking through.

Tested this on host via bionic-unit-tests-run-on-host on both x86 and
x86_64.

Change-Id: I2f06814e82c8faa732cb4f5648868dc0fd2e5fe4
Signed-off-by: Pavel Chupin <pavel.v.chupin@intel.com>