Make tests more tolerant of pthread test failures.
If pthread_key_create_lots fails after allocating all (or most) of its
keys, it is likely to cause many other tests to fail for the wrong
reasons. This test should probably be run in its own process, but for
now we can add a scope guard to protect against this.

Change-Id: I14af29a0a6869d8c12b486fd32c2439365289d0d