Fix linking static tests without libgcc.
The only reason this has been working so far is because I managed to
break a part of the build system when I introduced LOCAL_CXX_STL to
the build system. The build system is supposed to exclude libgcc when
using libc++ to ensure that we get the correct unwinder. I had broken
this check, so we were actually compiling with libgcc in this case.
Fixing this in the build system requires explicitly linking the
unwinder and compiler-rt.

Bug: 17924600
Change-Id: I3e92e3c9b90530afadae8121be0cd1409648adc4