Update patch set 1
Patch Set 1:

This patch is fine, but restricting this entirely to C11+ is tricky.  If my patch to move android_atomic to stdatomic.h ever makes it, we end up with a lot of stdatomic.h clients that don't know they are.  I played a little with moving some of those compilations to C11, but it seems nontrivial as you gathered.  It may be easier just to make stdatomic.h work in those contexts for now.

Patch-set: 1