Update patch set 1
Patch Set 1:

> > what is this testing that isn't already tested?
 > 
 > This adds the test for recursive type lock in cpp constructor.

but what's interesting about that? we don't have a test for pthread_mutex_lock in a function returning a double either, but no one would suggest we add one.

 > And
 > I don't see such test in pthread_test.cpp. It's the same as
 > pthread. pthread_mutex_timedlock or pthread_DeathTest.pthread_bug_37410
 > ?

the 37410 test relates to whether the thread list is corrupted when the main thread exits (which was once an actual bug). it also tests that pthread_join works for the main thread (which was also once broken) --- the main thread is a little special, so tests like this make sense.

but there's nothing obviously interesting about a C++ constructor though.

Patch-set: 1