Update patch set 3
Patch Set 3:

in this you could just load init.rc or the time zone data file or something. can we start with a more representative example of the kinds of things we want to test, so we can see what the makefile & script infrastructure actually needs to look like?

Patch-set: 3