Create system/etc/hosts if it doesn't exist.
This is a rather disgusting hack, but it seems the only alternative is
to require a top-level make before running the host tests. This isn't
a good alternative, because the buildbot running these tests shouldn't
need to build an entire tree (the tree won't be otherwise built
because this target will only be lunched for specifically these
tests).

Change-Id: Iaf9a5f53d0593a9e1a226e510553ad576cf37afb