Use the same unwinder as libc++.
We need to not export the libgcc unwinder, since some not completely
understood circumstances can cause the libgcc unwinder to call in to
the libc++ unwinder in other binaries, resulting in a segfault
(http://b/18317554). However, we can't simply stop exporting the
unwinder because of some backward compatibilty issues with the NDK, so
instead we just bundle the same unwinder that libc++ uses to ensure
compatibility. In the case of ARM, this is libunwind_llvm. For the
rest of the architectures, this is libunwindbacktrace. ARM also needs
libc++abi, because the EHABI unwinder depends on __gxx_personality_v0,
which is defined as a part of the C++ ABI, not the unwinder.

Bug: 18317554
Change-Id: Idb71b68e0a54c667d933d89087fd544566e3dc65