Update patch set 3
Patch Set 3:

Do I need to do anything with this build breakage? When I look into logs I can't understand how the breakage is related to my change. Do we have a list of all bionic tests which the bot checks? The error message is "Test reports were found but not all of them are new. Did all the tests run?"

Patch-set: 3