Update patch set 4
Patch Set 4:

We have no responsibility to test whether gtest is using pthread keys, right? And we also not intend to test whether we bionic is using pthread keys(as it is not a feature, not in posix standard). This test just take advantage of the fact that gtest is using pthread keys. It may used to test up boundary of pthread keys we can create.

I guess you may want to change bionic_tls.h to fix the total number of pthread_keys we can create in all to PTHREAD_KEYS_MAX. However, it is not the behavior as before. I am not sure about this.

Patch-set: 4