Update patch set 1
Patch Set 1:

can we just change st_size for mips to be off64_t? we should change arm64 and x86-64 to be off_t (which is what POSIX says). i already have a bug to go through and try to fix the types of those struct stat members we can fix without breaking the ABI but as you can see i've fixed relatively few of the members.

Patch-set: 1