Update patch set 1
Patch Set 1:

no, don't worry, you're not behind. there's just one build that's currently defaulting to clang instead of defaulting to gcc. you _can_ do that, but then you're not testing gcc, which is still the default for everyone :-)

what you did was right. test with gcc, fix clang issues as they come up. just next time (a) try to include the compiler's error message in the checkin comment of a change that just fixes a build and (b) say "TBR build fix" when self-approving.

Patch-set: 1