Update patch set 1
Patch Set 1: Code-Review-1

PTHREAD_STACK_MIN is a pain in the ass. it's not really useful to user code -- because it makes no guarantees about what you can do with your stack when we call you, only that it's enough for us to be able to set up a new thread and jump to the start_routine you passed us -- but developers do seem to interpret it as "hey, here's a handy value i can use for my worker threads".

if we do change it, i think we should probably just accept that it's always abused and just set it to 128KiB for both LP32 and LP64. (because that is a sensible size for a worker thread.)

right now, though, i think the actual bug is that we forgot to change the arithmetic in pthread_create. if they ask for an n-byte stack, we should give them n *usable* bytes. so we should allocate a region that's n + sizeof(our stuff).

Patch-set: 1
Label: Code-Review=-1