Update patch set 1
Patch Set 1:

Use of the -integrated-as switch is no longer needed with mips clang, and sometimes causes its own bugs.

With the Dec 22 refresh of prebuilts/gcc mips64el 4.9, gcc gas is now able to assemble the output of clang, without getting the warnings about $at register.  So the -integrated-as switch is no longer needed when using clang.

clang is now able to compile bionic for mips32r2, mips32r6, and mips32r6, with or without the -integrated-as switch.

Imgtec has recently discovered illegal-opcode problems with mips clang when compiling v8 with the -integrated-as switch.  Until we resolve that, we recommend not using -integrated-as for any mips clang compiles.

Patch-set: 1