Update libc logd writes for log drops
With the introduction of log drop logging for logd, the wire format
for logs is now changed.  Update libc to always report a drop count of
0.  It doesn't share the drop counter with liblog (a thread-local),
and the logs coming from libc should be low volume anyway.  The
accuracy of log drops will not be 100% (under-reported), but it's
better than nothing.

Change-Id: Ie60fcb6395e8205883d7ad3a4d7f409fca3822b2