WIP: Untested: Change pthread_rwlock to writer preference.
This is almost certainly a bad idea.  Although this may improve behavior of
some KitKat applications, it's likely to break some new apps.

I'm uploading this only to get an idea what the code would look like
and how much work it would be to make this configurable by the client.

Bug: 19109156

Change-Id: I406716f9318bcf4efcaa0b6357108dfd4efc3cf3