Update patch set 1
Patch Set 1:

btw, feel free to make this controlled by an environment variable. we did that for the dynamic linker debugging, and it's a lot more useful now (and never bitrots). (we cheated a little for the dynamic linker and only read the environment variable when we first start, which makes it very cheap but might not be suitable for DNS.)

Patch-set: 1