Update patch set 1
Patch Set 1:

(4 comments)

> (2 comments)
 > 
 > some TODOs here:
 > 1. currently based on 3.16.3 headers, will do another check with
 > the latest 3.18 header
 > 2. the cleaned headers may be different from previous versions due
 > to (a) formatting, such as line break, indentation; (b) errors. the
 > differences are listed in uapi.diff file, and i'm investigating the
 > differences now.
 > 3. it feels like current version is running slow than previous
 > version. will time it and measure the difference later

that's something we can fix later, in a different change. this script only gets run once or twice a year. slowness is only a problem when you have to work on fixing bugs in the script itself. (and that seems like a good motivation to add a new test rather than work on the whole tree :-) .)

 > 4. cleaning up the cpp.py script to remove obsolete lines

likewise.

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0