Put back inline definitions if using an old API.
All these inlines were turned in to out of line definitions in L.
There have been a number of bug reports and SO questions about people
targeting API level 21, trying to run on pre-21, and getting errors
about the linker not being able to find rand(3) (or others).

While this isn't strictly supported, we can put the inline definitions
back for the older API levels. It clutters up the default namespace a
bit, but that's sort of how things were anyway.

This also brings us a step closer to being able to just use the
current bionic headers for the NDK, rather than having many old
versions of them.

Change-Id: Ie010bc727d78d3742abc577c70f6578db2e68625