Update patch set 1
Patch Set 1:

(8 comments)

Aside from some nits, the changes look good to me.

I think we should probably enlarge the thread id and count fields in the 64 bit case.  The current code seems dubious.  But that's a separate CL.  And it visibly breaks compatibility in the PSHARED case, which is mildly nasty.  Everything that touches the same mutex would need to be recompiled. 

This strikes me as a fairly high risk change, due to the still hand-wavy correctness argument.  But I think it remains correct, and it's clearly simpler than the original.

Patch-set: 1