Update patch set 1
Patch Set 1:

(2 comments)

It looks to me like this is no more broken than the original.

The memory ordering guarantees provided by the implementation are probably too weak.  If we assume that Posix' intent was to ensure sequential consistency for race-free programs, then this fails to provide that.  I don't think anyone has ever thought through the memory ordering issues for Posix semaphores.  If they had, they might have come up with a different interface.  The fact that getvalue() is exposed makes things fairly expensive to do correctly.  We can discuss this in more detail if it matters.

Do we know if this is performance critical for any applications?  If not, I would change all the atomic operations, except initialization, to sequentially consistent operations, get rid of the fences, and be done with it.  The code will be simpler and correct, but slower.

If that's not an option, I would change getvalue as suggested below, and we can then wait until Posix gets its act together.  But I would be upset if they actually allowed that implementation, since it violates the intent behind the C and C++ memory model rules.

Patch-set: 1
Label: Code-Review=0