add explicit_memset and fix explicit_bzero with it
Clang, GCC and other compilers special-case standard C functions like
memset. Calls to memset will be optimized out if nothing uses the
result, defeating usage as an information sanitization mechanism.

OpenBSD provides explicit_bzero to work around this but Android simply
defines it as memset so nothing prevents it from being optimized away.

The explicit_memset function comes from NetBSD and can be used as a
drop-in replacement for memset with the guarantee that it won't be
optimized out. It is more sensible than starting from the legacy bzero
function. Android is currently misusing memset for sanitization in
various places such as zeroing the initial random seed in the init
process and should be migrated over to this function.

This implementation uses a memory read constraint via empty inline
assembly rather than something that may be broken via link-time
optimization in the future. The compiler will still happily expand the
memset call to some optimized inline code, but is forced to keep the
whole write operation intact if the memory isn't guaranteed to hold that
value already.

Change-Id: I2d405aa90f7998510015206e1e04b0e2f868f64e
Signed-off-by: Daniel Micay <danielmicay@gmail.com>