Update patch set 1
Patch Set 1:

(1 comment)

I'm confused about what this is trying to do.  Add more of an explanation in the code?

Are you trying to measure the overhead of a futex_wake of a waiting thread?  If so, would it work to set up a cycle of threads, each one of which waits for its predecessor to wake it and then wakes its successor?

Patch-set: 1
Label: Code-Review=0