Update patch set 1
Patch Set 1: -Code-Review

We are trying to measure the overhead cost in the producer for futex_wake in the context of a sem_post. We do _not_ want to measure the cost in the consumer of the sem_wake since that is highly variable and subject to scheduling. The variability is still somewhat noisy though since the futex wake call can vary because of CPU clock or ancillary overhead in the system call to evaluate which thread is to run next if at all before return (hence why producer is at SCHED_FIFO).

Good catch on the loss of the __sem_wake_run ... I must have missed it in the forward-port to AOSP from the test systems in goog/master <eek>

Patch-set: 1