Update patch set 1
Patch Set 1:

After a deeper look, it appears that the only place where we play with 2 locks is pthread_key_delete(). If we add g_thread_list_lock inside ScopedTlsMapAccess lock mgt, we will have the risk of double lock (AA). So, the best thing to do is to define a lock hierarchy: g_thread_list_lock before s_tls_map_lock_. That's the strategy of the current patch.

Patch-set: 1