Update patch set 1
Patch Set 1: -Code-Review

(1 comment)

What you speak of regarding go-fast should not happen if the semaphores are working in lock step and at SCHED_OTHER policy; although I worry about two processors running simultaneously. If they did, the result would end up being small as compared to BM_semaphore_sem_post. One would have to be 'alarmed' when they see that result meaning something is amiss.

I think the only way I can make this test guarantee repeatable and not get into a go-fast mode in the 2-thread case is to also lock the primary and background thread to the same core. I can do affinity in this special case, we can always assume CPU0 is running to KISS.

Otherwise the problem is on a N6 for example, with this test, and with the interactive governor, a test with 3 or more (which would _not_ go-fast) would already report a degraded performance from not having enough processors active. So I can not use 3 or 4 or so as the BM_semaphore_sem_psot_sem_wait_low test.

Another way I can ensure lock step is switch to SCHED_IDLE for the main and spawned threads as they will not be allowed to run until nothing else is requiring the system. It will make them highly susceptible to anything happening on the system though. I will investigate this alternative as well.

Patch-set: 1