Update patch set 1
Patch Set 1:

(1 comment)

an alternative argument is "it's not valid to implement strnlen in terms of memchr". which would give us two choices:

1. also switch to OpenBSD's strnlen that doesn't have this optimization.
2. leave memchr and memrchr alone and only switch strnlen.

another argument is "these size_t arguments must be less than the remaining size of the allocation that the correponding pointer points into", in which case we could just say this bug is WAI.

thoughts?

Patch-set: 1