Update patch set 10
Patch Set 10:

Looking at the code, I don't see any way that any floating point exception could occur. The only instruction that could generate one is the vsub. However, it looks to me like both operands are constrained such that it's not possible to generate an exception.

Is that the exception handling you are looking for, or is there something else?

Patch-set: 10