Update patch set 2
Patch Set 2:

> AFAICT, a fairly typical use is to look up a thread-local
 > allocation buffer in a memory allocator like malloc.  Since each
 > call allocates a single object, there is no place to stash the
 > value obtained from pthread_getspecific() between calls.  In fact,
 > pthread_getspecific() is effectively being used to stash allocation
 > buffer pointer between calls.
 > 
 > My own experience with this in a memory allocator several years ago
 > was that the standard Linux/glibc implementation was too slow to be
 > really useful.

it's roughly the same speed as our current implementation on x86-64. jemalloc uses it to choose an arena so it can't be too bad. in art we started off with no caching but found that it was a worthwhile saving to cache in hotspots where we could reuse it.

 > I ended up implementing my own hack that was
 > basically a map of high order stack bits to the appropriate value. 
 > It may be that modern code prefers thread_local or __thread, so
 > this is less critical.

we don't have native support for them, and the fallback implementation is pthread_getspecific/pthread_setspecific :-)

Patch-set: 2
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0