Update patch set 1
Patch Set 1:

> I'm not aware of any environment variable used by the core
 > init.environ.rc file which would be inappropriately filtered by
 > this. There may be device-specific environment variables, but if
 > so, they can add an allow rule in their device-specific policy.
 > Setting these kinds of environment variables from init feels
 > inappropriate to me.
 > 
 > At some point, I want to revisit init being able to load new
 > security policy. It's been one of the top concerns I've heard from
 > partners, since they've carefully tested their devices and are
 > worried about something mucking with the policy. It's actively used
 > by superuser type apps to muck with policy. It's used as an excuse
 > to not tighten up SELinux policy. And the update process is a
 > poorly tested code path that we've never used. I understand that
 > policy updates could be used to mitigate vulnerabilities in the
 > field, but I'm not convinced that advantage outweighs the
 > downsides.

I was thinking that it was more likely to be used for quick policy updates to allow something to untrusted_app that they didn't envision or test for prior to shipping, as it is problematic to test the universe of apps.  In other words, to fix app breakage / avoid incompatibilities.  Otherwise, their only option is to ship an OTA with an updated policy for each such fix which I had the impression is harder to get deployed particularly via certain carriers.
For our usage scenarios, it allows custom policy without having to maintain a custom ROM, as the latter is not viewed as viable for enterprise deployments.  Not too concerned with rooted devices as that presumes that you've already unlocked the bootloader and all bets are off at that point.  Even without SELinux load_policy, init is necessarily the most powerful userspace process on the device.

Patch-set: 1