Loose fchmodat AT_SYMLINK_NOFOLLOW test on symlink.
It has been reported in b2/19657449 and b2/19381040 that fchmodat
AT_SYMLINK_NOFOLLOW operation on symlink can succeed. It seems to be
controlled by kernel or user configuration whether chmod is allowed
on symlinks. Unless we can disable chmod on symlinks in bionic explicitly,
we can not guarantee that the test can pass. So I prefer to loose the
test here.

Bug: 19657449
Bug: 19381040
Change-Id: I780e84f0b50d0412fbac9f1c240d07e984892a28