linker : Temporary Workaround Fix for grouper gpu blobs
The Nvidia blobs for Grouper fails to set the correct working
directory when loading the OpenGLES libraries using relative paths
Accomadating this in the linker causes the incorrect behaviour
described in bug 6670.

I expect this workaround to be very temporary until some fixed
proprietary binaries are recompiled or the grouper is no more.
In meantime this workaround lets the grouper play nice with the
way dlopen is expected to function.

https://code.google.com/p/android/issues/detail?id=6670

Change-Id: I7f1f88e6da5260df91b91f7fad3acf28024ec162